+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added action link.
+- Added `back_link` to fixed elements 
 
 ## [0.11.0] - 2026-01-16